Fix build, style
